refactor: introduce matching package in core
