+++ pkg/core/matching/number.go
@@ -1,4 +1,4 @@
-package reflect
+package matching
 
 import (
 	"reflect"
