+++ pkg/core/matching/number_test.go
@@ -1,4 +1,4 @@
-package reflect
+package matching
 
 import (
 	"reflect"
